requirements.txt
Django==3.0.5
django-admin-rangefilter==0.5.4
pytest==5.4.1
pytest-django==3.9.0
djangorestframework==3.11.0
psycopg2-binary==2.8.3